+++ R/messages.R
@@ -161,3 +161,19 @@ messages$errorWrongArguments <- function(expectedArguments) {
     expectedArguments, "."
   )
 }
+
+messages$invalidConfigurationPropertyFromExcel <- function(propertyName, configurationType) {
+  paste0("Trying to apply property '", propertyName, "' that is not supported by
+         the configuration '", configurationType, "'! Check column names in the
+         excel file defining plot configurations.")
+}
+
+messages$missingOutputFileName <- function() {
+  paste0("Missing values found in mandatory column 'outputName' of sheet 'exportConfiguration'. No plots are exported to file for corresponding rows.")
+}
+
+messages$missingPlotGrids <- function(missingPlotGrids) {
+  paste0("Invalid values in column 'plotGridName' of sheet 'exportConfiguration': ",
+         paste(missingPlotGrids, collapse = ", "),
+         ". Plot grids are either not defined or empty and can not be exported to file.")
+}